Move widgets into core/widgets dir
