+++ hamsa_lomi/lib/presentation/home/page/home_page.dart
@@ -2,7 +2,7 @@
 import 'package:flutter/material.dart';
 
 // Project imports:
-import '../../core/hamsa_app_bar.dart';
+import '../../core/widgets/hamsa_app_bar.dart';
 import '../widgets/custom_bottom_nav_bar.dart';
 import '../widgets/fundraising_filter.dart';
 import '../widgets/start_fundraiser_button.dart';
