+++ hamsa_lomi/lib/presentation/lang_acct_type/pages/choose_account_type.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 // Project imports:
 import '../../constants/app_assets_constant.dart';
 import '../../constants/app_string_constant.dart';
-import '../../core/hamsa_app_bar.dart';
+import '../../core/widgets/hamsa_app_bar.dart';
 import '../../create_account/pages/create_account_page.dart';
 import '../../theme/hamsa_theme.dart';
 import '../widget/button_widget.dart';
